+++ CHANGELOG.md
@@ -1,6 +1,22 @@
+## 0.4.0
+
+- `noBackground` prop is added, if given `Wowerlay` won't have an invisible click blocker background when visible.
+
+- `scrollbarGap` not counted for `checkOutOfScreen` is fixed.
+
+- Readme is updated.
+
+### Breaking Changes
+
+- `WowerlayContainer` component is removed, now `Wowerlay` comes with its own background.
+
+<br>
+
 ## 0.3.2
 
 - Updated README.md
+
+<br>
 
 ## 0.3.1
 
@@ -9,6 +25,8 @@
 ### Demo
 
 - `With Gaps` demo is added.
+
+<br>
 
 ## 0.3.0
 
@@ -26,11 +44,15 @@
 
 - `centered` prop is removed, now `top` | `left` | `right` | `bottom` is centered by default.
 
+<br>
+
 ## 0.1.1
 
 - Now Wowerlay will close child Wowerlays on click itself.
 
 - `centered` prop added, if true wowerlay will be centered.
+
+<br>
 
 ## 0.1.0
 
@@ -43,6 +65,8 @@
 - Reactive `target` prop support is added. Now wowerlay will react when `target` element changes and will follow it.
 
 - Demo page added to Readme
+
+<br>
 
 ## 0.0.1